backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==0.21.0
sentence-transformers==2.2.2
numpy==1.24.0
scikit-learn==1.3.0
python-magic==0.4.27
aiofiles==23.2.1
openai==1.3.0
httpx==0.25.0